requirements.txt
Flask==2.2.2
requests==2.28.1
gunicorn==20.1.0
PyGithub==1.55.0
PyMuPDF==1.19.6